Make the module namespace ID a large fixed integer
Fix for bug 37641. Referencing a constant defined in the extension
registration file breaks some scripts. Referencing a constant defined in
an extension setup function is even worse. So it's not really possible
to configure the namespace number.

To update:

UPDATE page SET page_namespace=page_namespace+808
WHERE page_namespace IN (20,21);

Change-Id: I18e53127147f62a8b2058537f7b054c253ba126e