Lua ustring implementation
This is a reimplementation of Lua's string module with support for
UTF-8.

In LuaStandalone, the library is implemented in pure Lua.

In LuaSandbox, the library is implemented using PHP callbacks, which
particularly speeds up the pattern matching functions since the patterns
can be translated to PCRE and run that way.

It also includes a bunch of unit tests.

Note that if you download and run the Unicode Normalization Test Suite,
98 of the test strings currently fail. But MediaWiki core's UtfNormal
(the version used if the php intl and icu extensions are not available)
fails those same 98, which means it's probably a problem in the data
tables shared by both implementations.

Change-Id: Ie76fdf8d3a85d0a3d2a41b0d3b7afe433f247af0