Remove long-obsolete ScribuntoEngineBase::getLimitsReport()
In 925045a66 the method was renamed from getLimitsReport() to
getLimitReport(), but this instance was overlooked. Since every existing
subclass implements getLimitReport(), that wasn't noticed.

Therefore, when Idffd2d78 deprecated getLimitReport() it too missed this
instance (and, in fact, added its own copy of a base-class
getLimitReport() so it could mark that deprecated instead), and
I0e8e1d56 similarly missed removing it.

So let's kill it now, like it should have been long ago.

Change-Id: I9aa71b4a90b9e7cdda08ce171b3bb004c8ccd8c2