LuaStandalone: Improve passing of numeric values
For integers from Lua to PHP, make sure they won't use exponential
notation that will confuse unserialize(), and pass the integer size from
PHP so Lua can know which numbers are representable as integers.

For doubles in both directions, increase the precision to avoid
truncation of the least significant bits.

Change-Id: Icfaff71cab0ee1aac04acf752d108049b5569380