tests: Mark non-LuaSandbox tests as @group Standalone
This means that these tests are only run for test runs on this repo,
and will not be run when this repo is included in CI as a dependency
of any other repos, or the Wikimedia-wide gate. Any changes to other
repos are implicitly assumed to be unable to break these tests.

Bug: T125050
Change-Id: I7b77c0627f3fbe9dfbbf3d63cbd57e943d328555