Temporary debug logging for LuaEngine::frameExists('current')
~1.6% of PHP wall time on the Wikimedia cluster is burnt on calls to
Scribunto_LuaSandboxCallback::frameExists() from Lua (php.frameExists).
On a dev instance of MediaWiki, most calls are checking if
frameExists('current'), which I think must always be true. This logging
code is not a substitute for complete case analysis, but it's a cheap
way to get some validation before I burn more time staring at this code.

If I'm very wrong and there's lots of logspam, I'll revert this.

Change-Id: I84e8732d8d1e4b24c1b6eea3a9119d725c971c9a