Improve error messages about unknown code location
This is motivated by the WIT team's work on T380291. Both messages
just confused us a lot while working with this code. As it turns out
the previous "[UNKNOWN]" means something specific: that there is no
stack trace. Same for the vague mention of "further details" in the
user-facing message. That's where the stack trace usually renders
(via the message "scribunto-common-backtrace" above).

Bug: T380291
Change-Id: Ic5dbd0eda84a8421c01714ce9079c0dd15af1652