Allow adding multiple warnings using mw.addWarning
Specify an incrementing deduplication key when adding the warning
message to the ParserOutput.
Before this change, the key of the message, 'scribunto-lua-warning',
was used as the deduplication key, so if a second warning was added,
it would override the first one.

Also add a test to verify that the new behaviour works properly.

Bug: T398390
Change-Id: I6a94cba7cef97e488583de8cf3f40d5c52fa01a0