Cache the default engine in EngineFactory::getDefaultEngine
Remove ScribuntoEngineBase::setTitle and pass the title as extra
argument to EngineFactory::getDefaultEngine. This avoids that the
ScribuntoEngineBase can be changed from outside of the class and
allows to cache this engine instance in the factory,
when no extra arguments are used.
Within the parse operation of a wikitext, the same engine is used as
well, so it should be safe, when used outside of the parser.

The cached default engine can be used by some hook handlers and
for redirect contents.

Rename EngineFactory::newDefaultEngine to
EngineFactory::getDefaultEngine to reflect,
that this may reuse a cached engine and not create a new engine.

Change-Id: I3ab121c7131a3243e89361b0289216ead4d86ff0