Remove use of Parser::$scribunto_engine
Replace caching of engine on the Parser object with
a WeakMap in the factory and store the engine under the parser object.

hooks:
Remove the ParserCloned hook, it is no longer needed.
On clone the old property was copy over to the new instance and this
hook has destroyed the reference to the old engine.
But the fresh parser does not has a engine in the factory cache.

tests:
Remove use of the property, as no reference to the parser exists to
reuse the engine via hooks.

Bug: T343229
Change-Id: Ib9d96bea63f5157ed80f309a66d12de8f2e25c0c