mw.language: Add mode argument to getFallbacksFor()
This makes it possible for Lua code to ask for strict fallbacks, rather
than message fallbacks, i.e. without the implicit fallback to 'en' at
the end (but *with* any explicit fallback to 'en', if the language has
them). We expose the mode arguments to Lua as custom strings which
LanguageLibrary internally maps back to the MediaWiki constants, so the
MediaWiki constants can change (potentially become enums) without
breaking anything on the Lua side.

The error handling for invalid modes – (ab)use the “one of” string as
the expected type in the error message – is borrowed from the $which
argument of SiteLibrary::pagesInCategory().

In the tests, test_multi needs a small tweak so that it will result in
calls like

    mw.language.getFallbacksFor( 'en' )
    mw.language.getFallbacksFor( 'de' )
    mw.language.getFallbacksFor( 'arz' )
    mw.language.getFallbacksFor( '[[bogus]]' )

rather than

    mw.language.getFallbacksFor( 'en', 'de', 'arz', '[[bogus]]' )
    mw.language.getFallbacksFor( 'de', 'arz', '[[bogus]]' )
    mw.language.getFallbacksFor( 'arz', '[[bogus]]' )
    mw.language.getFallbacksFor( '[[bogus]]' )

(There might be a more elegant way to do this, I’m not very familiar
with Lua varargs / select().)

Bug: T408998
Change-Id: Ifb43718ec2a8e939d62c7f54f1eee762a84848bc